Started working on the editing feature
When pressing e it'll simply open up a temporary file with the values
in your favorite editor (or vi if none is set). Idea is to after this
file is closed in the editor to parse it and set the values to the nbt
data and all. That however isn't done yet atm.

Signed-off-by: Toon Schoenmakers <nighteyes1993@gmail.com>