We're now able to save our data back to the original file
Also fixed a tiny bug in the error() method where it apparently didn't
use the arguments. Next to this made an exact copy called notice() which
does exactly the same but in green, so mainly to easily confirm stuff.
It also appears that the library doesn't save the data in the same order,
meaning that most of the times even if you didn't change anything at all
the output file will still be different. But this seems to be something
in the cNBT library which I am not going to fix.

Signed-off-by: Toon Schoenmakers <nighteyes1993@gmail.com>